pyproject.toml
[project]
name = "opensymbolicai-cli"
version = "0.1.0"
description = "CLI for OpenSymbolicAI - Interactive TUI for discovering and running agents"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
    "textual>=0.89.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "opensymbolicai-core",
    "httpx>=0.28.1",
]

[project.scripts]
opensymbolicai = "opensymbolicai_cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "build>=1.0",
    "mypy>=1.0",
    "pre-commit>=4.0",
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "python-semantic-release>=10.5.3",
    "ruff>=0.9",
    "twine>=6.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
no_implicit_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
branch = "main"
commit_message = "chore(release): {version}"
build_command = "uv run python -m build"
tag_format = "v{version}"

[tool.semantic_release.remote]
type = "github"
name = "origin"

[tool.semantic_release.publish]
upload_to_vcs_release = true
